+++ 1.21.10/common/src/main/java/com/iamkaf/mochila/registry/Items.java
@@ -114,6 +114,7 @@ public class Items {
     }
 
     public static void init() {
+        CreativeModeTabs.addItem(ENDER_BACKPACK);
         ITEMS.register();
     }
 }